Put message API in its own file
